[Friend] Restructure my lovely reducers
