+++ src/App/ActionTypes.js
@@ -14,6 +14,8 @@ export const BEER_FETCH_SUCCESS = 'BEERS_FETCH_SUCCESS';
 export const BEER_UPDATE = 'BEER_UPDATE';
 
 // Friend
+export const FRIEND_ADD_DIALOG_CLOSE = 'FRIEND_ADD_DIALOG_CLOSE';
+export const FRIEND_ADD_DIALOG_OPEN = 'FRIEND_ADD_DIALOG_OPEN';
 export const FRIEND_FETCH = 'FRIEND_FETCH';
 export const FRIEND_FETCH_SUCCESS = 'FRIEND_FETCH_SUCCESS';
 export const FRIEND_SEARCH_INFO_UPDATE = 'FRIEND_SEARCH_INFO_UPDATE';
