+++ src/Friend/actions/FriendSearchAction.js
@@ -1,0 +1,14 @@
+// External Dependencies
+// import firebase from 'firebase';
+
+// Internal Dependencies
+import {
+  FRIEND_SEARCH_INFO_UPDATE,
+} from '../../App/ActionTypes';
+
+export const friendSearch = () => {};
+
+export const friendSearchInfoUpdate = ({ prop, value }) => ({
+  type: FRIEND_SEARCH_INFO_UPDATE,
+  payload: { prop, value },
+});
